add fix with test coverage

+++ Domain.Sql.Tests/ReportingProjectorTests.cs
@@ -286,7 +286,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
             var expectedAggregateId = Any.Guid();
             var receivedAggregateId = new Guid();
             var expectedSequenceNumber = Any.PositiveInt();
-            int receivedSequenceNumber = 0;
+            long receivedSequenceNumber = 0;
             var receivedTimestamp = new DateTimeOffset();
             var receivedEventStreamName = "";
             var receivedEventTypeName = "";
@@ -611,7 +611,7 @@ namespace Microsoft.Its.Domain.Sql.Tests
         public int Quacks { get; set; }
     }
 
-    public class Reporting
+    public class  Reporting
     {
         public class DuckEvent
         {
